GH-7 Replace Geocoding library
The existing Geocoding library was too limited, it only provided
latitude and longitude when the Google Geocoding API provides much
more.

The library is replaced with a model class to hold the JSON
response from the API query, as well as using the formatted_address
field to return a well formatted location string for the response.